Reliable paged comments
This commit uses FQL to get very reliable
comment system, where you can load pages
of comments and be sure that this will work.

One only bug is that if your number of comments
is multiple of ten, then the "load previous"
button is displayed even if there are no more
comments.